refactor(vscode): extract bundler launcher into a testable module
The launcher logic — command shims, missing-executable detection,
shell quoting, candidate fallback, and command resolution — lived as
private functions inside extension.ts, untestable without loading
the VS Code API. Move it to bundlerLauncher.ts with no vscode
import: the extension reads configuration and passes plain options,
and runBundler accepts an injectable candidate runner that defaults
to the real execFile path, so production behavior is unchanged
while tests can fake spawning.
